[IMAGING-159] Remove Imaging class methods that accept parameters (users must use the ImageParsers directly)

+++ src/main/java/org/apache/commons/imaging/formats/tiff/write/TiffImageWriterBase.java
@@ -341,7 +341,7 @@ public abstract class TiffImageWriterBase {
         int stripSizeInBits = 64000; // the default from legacy implementation
         Integer compressionParameter = params.getCompression();
         if (compressionParameter != null) {
-            compression = compressionParameter.intValue();
+            compression = compressionParameter;
             final Integer stripSizeInBytes = params.getLzwCompressionBlockSize();
             if (stripSizeInBytes != null) {
                 if (stripSizeInBytes < 8000) {
